% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucaspkg.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{The log summed exponential function}
\usage{
log_summed_exps(n, x = seq(1:n))
}
\arguments{
\item{n}{A positive integer value indicating the sample size}

\item{x}{A vector of numeric values (set to i, sequence of positive integers
from 1 to the sample size n, as default)}
}
\value{
The log summed exponential of the input vector
}
\description{
The log summed exponential function
}
\examples{
# basic usage of log_summed_exps
log_summed_exps(10)

# elaborate usage of log_summed_exps
log_summed_exps(5, c(1,43,6,54,123))

}
