% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Healy.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{The transformation of log(sum(e^2))
using the fact log(x+y)=log(x(1+y)/x)}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
The log of summation of e^x
}
\description{
The transformation of log(sum(e^2))
using the fact log(x+y)=log(x(1+y)/x)
}
\examples{
## Pass a numeric vector to log_summed_exps,
## and it will log the summation of e^x
log_summed_exps(seq(1, 10))

}
