% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Yimin's hw}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of numeric}
}
\value{
log of summed exponential
}
\description{
Yimin's hw
}
\examples{
library(YiminZhao)
test = seq(1, 2000, 1)
log_summed_exps(test) # expected to get 2000.459
}
